Fix error handling
Move clean() function from Optimization to OptimizationManager. Extend Try block
